rework contributing to outline all scripts, remove postinstall

+++ rsbuild.config.ts
@@ -101,6 +101,9 @@ export default defineConfig({
                     if (!fs.existsSync('./generated/minecraft-data-data.js')) {
                         childProcess.execSync('tsx ./scripts/genShims.ts', { stdio: 'inherit' })
                     }
+                    if (!fs.existsSync('./generated/latestBlockCollisionsShapes.json')) {
+                        childProcess.execSync('tsx ./scripts/optimizeBlockCollisions.ts', { stdio: 'inherit' })
+                    }
                     fsExtra.copySync('./node_modules/mc-assets/dist/other-textures/latest/entity', './dist/textures/entity')
                     fsExtra.copySync('./assets/background', './dist/background')
                     fs.copyFileSync('./assets/favicon.png', './dist/favicon.png')
@@ -115,7 +118,7 @@ export default defineConfig({
                     // childProcess.execSync('./scripts/prepareSounds.mjs', { stdio: 'inherit' })
                     // childProcess.execSync('tsx ./scripts/genMcDataTypes.ts', { stdio: 'inherit' })
                     // childProcess.execSync('tsx ./scripts/genPixelartTypes.ts', { stdio: 'inherit' })
-                    if (fs.existsSync('./prismarine-viewer/public/mesher.js')) {
+                    if (fs.existsSync('./prismarine-viewer/public/mesher.js') && dev) {
                         // copy mesher
                         fs.copyFileSync('./prismarine-viewer/public/mesher.js', './dist/mesher.js')
                     } else {
